backend/requirements.txt
flask
flask-cors
requests
gunicorn
sqlalchemy
google-generativeai>=0.7.2
python-dotenv
